FUSETOOLS2-833 - use waitUntil to avoid flakiness in test
